Migrate rating and notes section from GameInfo to new comp
